+++ src/lib/pages/code-details/components/code-info/index.ts
@@ -1,2 +1,3 @@
-export * from "./table/CodeContractsTable";
+export * from "./table/CodeContractsTableFull";
+export * from "./table/CodeContractsTableLite";
 export * from "./CodeInfoSection";